pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "Cython", "setuptools-scm", "tomli", "scipy"]
build-backend = "setuptools.build_meta"

[project]
name = "pgam_clean"
dynamic = ["version"]
authors = [{name = "nemos authors"}]
description = "NEural MOdelS, a statistical modeling framework for neuroscience."
readme = "README.md"
requires-python = ">=3.12,<3.14"
keywords = ["neuroscience", "PGAM"]
license = { file = "LICENSE" }
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]

# Define dependencies for the project
dependencies = [
    "nemos",
    "matplotlib",
    "jaxopt",
    "optimistix<0.1.0",
    # only needed for comparrison with origina implemetation
    "PGAM",
]


[tool.uv.sources]
PGAM = { path = "../PGAM", editable = true }

# Configure package discovery for setuptools
[tool.setuptools.packages.find]
where = ["src"]     # The directory where package modules are located
include = ["pgam_clean"] # The specific package(s) to include in the distribution

[tool.setuptools_scm]
version_scheme = 'python-simplified-semver'
local_scheme = 'no-local-version'

# Define optional dependencies for the project
[project.optional-dependencies]
dev = [
    "black",                        # Code formatter
    "isort",                        # Import sorter
    "pip-tools",                    # Dependency management
    "ruff",                         # Lint docstrings for now, eventually switch to ruff for all linting
    "pytest",                       # Testing framework
    "pytest-xdist",                 # Parallelize pytest
    "flake8",                       # Code linter
    "coverage",                     # Test coverage measurement
    "pytest-cov",                   # Test coverage plugin for pytest
    "statsmodels",                  # Used to compare model pseudo-r2 in testing
    "scikit-learn",                 # Testing compatibility with CV & pipelines
    "matplotlib>=3.7",              # Needed by doctest to run docstrings examples
    "seaborn",                      # Required by doctest for _documentation_utils module
    "myst-nb",                      # Test myst_nb utils for glue
    "tox-uv",                       # Run tox with uv for faster installs
    "jupyterlab",
    "jupyterlab-vim",
    "jupyterlab-code-formatter",
]
docs = [
    "numpydoc",
    "sphinx",
    "pydata-sphinx-theme",
    "sphinx-autodoc-typehints",
    "sphinx-copybutton",
    "sphinx-design",
    "sphinx-issues",
    "sphinxcontrib-apidoc",
    "sphinxcontrib-bibtex",
    "sphinx-togglebutton",
    "sphinx_code_tabs",
    "sphinxemoji",
    "myst-parser",
    "myst-nb",
    "dandi",
    "sphinx-autobuild",
    "sphinx-contributors",
    "scikit-learn",
    "dandi",
    "matplotlib>=3.7",
    "seaborn",
    "pooch",
    "ipywidgets",
    "ipykernel",
    "patsy",
    "tqdm",
]
examples = [
    "scikit-learn",
    "dandi",
    "matplotlib>=3.7",
    "seaborn",
    "pooch",
    "fsspec",
    "patsy",
    "tqdm",
]
jaxopt = [
    "jaxopt>=0.6",
]


[tool.black]
target-version = ['py310', 'py311', 'py312', 'py313']
skip-string-normalization = false
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.md
    | \.toml
    | \.cfg
    | \.txt
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | examples
    ))'''

# Configure isort
[tool.isort]
multi_line_output = 3             # Use three-line style for multi-line imports
include_trailing_comma = true     # Include trailing comma in multi-line imports
profile = "black"
skip = []

# Configure pytest
[tool.pytest.ini_options]
testpaths = ["tests"]             # Specify the directory where test files are located
filterwarnings = [
    # note the use of single quote below to denote "raw" strings in TOML
    # this is raised whenever one imports the plotting utils
    'ignore:plotting functions contained within:UserWarning',
    # numerical inversion test reaches tolerance...
    'ignore:Tolerance of -?\d\.\d+e-\d\d reached:RuntimeWarning',
    # mpl must be non-interctive for testing otherwise doctests will freeze
    'ignore:FigureCanvasAgg is non-interactive, and thus cannot be shown:UserWarning',
]
markers = [
  "solver_related: tests that interact with solvers",
  "requires_x64: mark test as requiring JAX float64 precision",
]
# -s: no capture from pytest (prints will show)
addopts = ["-s", "--strict-markers"]

[tool.coverage.run]
omit = [
    "src/nemos/fetch/*",
    "src/nemos/_documentation_utils/*",
]

[tool.coverage.report]
exclude_lines = [
    "@abc.abstractmethod",
    "if __name__ == .__main__.:"
]


[tool.ruff]
exclude = ["src/nemos/_documentation_utils"]


[tool.ruff.lint]
# TODO: Re-enable NumPy docstring rules when public APIs are ready.
# select = ["D"]
# ignore = ["D202"]  # this conflicts with black
select = []


[tool.ruff.lint.pydocstyle]
convention = "numpy"
